requirements.txt
# Add your dependencies here
sentence-transformers
qdrant-client
feedparser
numpy
streamlit
networkx
pyvis
google-generativeai
python-dotenv